- adds routeUtils to simplify common imports - adds getDefaultMiddleware to simplify common middleware imports
