+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "node-starter-app",
-  "version": "0.0.4",
+  "version": "0.0.5",
   "description": "",
   "author": "CDS",
   "license": "MIT",
