+++ routes/_sample_route/sample.controller.js
@@ -1,26 +1,18 @@
 const path = require("path");
-const { checkSchema } = require("express-validator");
-const {
-  addViewPath,
-  getRouteByName,
-  checkErrors,
-  getSessionData,
-  doRedirect
-} = require("./../../utils");
+const { routeUtils } = require("./../../utils");
 const { Schema } = require("./schema.js");
 
 module.exports = app => {
   const name = "sample";
-  const route = getRouteByName(name);
+  const route = routeUtils.getRouteByName(name);
 
-  addViewPath(app, path.join(__dirname, "./"));
+  routeUtils.addViewPath(app, path.join(__dirname, "./"));
 
-  app.get(route.path, (req, res) => {
-    res.render(name, {
-      data: getSessionData(req),
-      name
-    });
-  });
-
-  app.post(route.path, checkSchema(Schema), checkErrors(name), doRedirect);
+  app
+    .get(route.path, (req, res) => {
+      res.render(name, routeUtils.getViewData(req, name));
+    })
+    .post(route.path, [
+      ...routeUtils.getDefaultMiddleware({ schema: Schema, name: name })
+    ]);
 };
