+++ utils/data.helpers.js
@@ -1,0 +1,23 @@
+const { getSessionData } = require("./session.helpers");
+const { getFlashMessage } = require("./flash.message.helpers");
+const { generateNonce } = require("./validate.helpers");
+
+const getViewData = (req, name) => {
+  const params = {
+    data: getSessionData(req),
+    name,
+    nonce: generateNonce()
+  };
+
+  const errors = getFlashMessage(req);
+
+  if (errors) {
+    params.errors = errors;
+  }
+
+  return params;
+};
+
+module.exports = {
+  getViewData
+};
